feat: #WB2-2701, fix i18n and session broker issues
