--- broker-parent/broker-client/ts/src/nats/directory-group-find-byexternalid.types.ts
@@ -1,13 +1,0 @@
-export interface FindGroupByExternalIdRequestDTO {
-  externalId?: string;
-}
-
-
-export interface FindGroupByExternalIdResponseDTO {
-  group?: GroupDTO;
-}
-export interface GroupDTO {
-  id?: string;
-  name?: string;
-}
-
