--- broker-parent/broker-client/ts/src/nats/directory-group-manual-create.types.ts
@@ -1,12 +1,0 @@
-export interface CreateGroupRequestDTO {
-  externalId?: string;
-  name?: string;
-  classId?: string;
-  structureId?: string;
-}
-
-
-export interface CreateGroupResponseDTO {
-  id?: string;
-}
-
