--- broker-parent/broker-client/ts/src/nats/directory-group-manual-delete.types.ts
@@ -1,10 +1,0 @@
-export interface DeleteGroupRequestDTO {
-  id?: string;
-  externalId?: string;
-}
-
-
-export interface DeleteGroupResponseDTO {
-  deleted?: boolean;
-}
-
