--- broker-parent/broker-client/ts/src/nats/directory-group-manual-update.types.ts
@@ -1,11 +1,0 @@
-export interface UpdateGroupRequestDTO {
-  id?: string;
-  externalId?: string;
-  name?: string;
-}
-
-
-export interface UpdateGroupResponseDTO {
-  updated?: boolean;
-}
-
