--- broker-parent/broker-client/ts/src/nats/directory-group-member-delete.types.ts
@@ -1,11 +1,0 @@
-export interface RemoveGroupMemberRequestDTO {
-  groupId?: string;
-  groupExternalId?: string;
-  userId?: string;
-}
-
-
-export interface RemoveGroupMemberResponseDTO {
-  removed?: boolean;
-}
-
