--- broker-parent/broker-client/ts/src/nats/share-group-remove-{application}.types.ts
@@ -1,19 +1,0 @@
-export interface RemoveGroupSharesRequestDTO {
-  currentUserId?: string;
-  groupId?: string;
-  groupExternalId?: string;
-  resourceId?: string;
-  application?: string;
-}
-
-
-export interface RemoveGroupSharesResponseDTO {
-  shares?: OrgEntcoreBrokerApiDtoSharesSharesResponseDTO[];
-}
-export interface OrgEntcoreBrokerApiDtoSharesSharesResponseDTO {
-  id?: string;
-  kind?: OrgEntcoreBrokerApiDtoSharesSharesResponseDTOKind;
-  permissions?: string[];
-}
-export interface OrgEntcoreBrokerApiDtoSharesSharesResponseDTOKind {}
-
