--- broker-parent/broker-client/ts/src/nats/{application}-i18n-translations-fetch.types.ts
@@ -1,19 +1,0 @@
-export interface FetchTranslationsRequestDTO {
-  headers?: {
-    [k: string]: string;
-  };
-  langAndDomain?: LangAndDomain;
-  application?: string;
-}
-export interface LangAndDomain {
-  lang?: string;
-  domain?: string;
-}
-
-
-export interface FetchTranslationsResponseDTO {
-  translations?: {
-    [k: string]: string;
-  };
-}
-
