--- broker-parent/broker-client/ts/src/nats/{application}-i18n-translations-register.types.ts
@@ -1,16 +1,0 @@
-export interface RegisterI18NFilesRequestDTO {
-  application?: string;
-  translationsByLanguage?: {
-    [k: string]: {
-      [k: string]: string;
-    };
-  };
-}
-
-
-export interface RegisterI18NFilesResponseDTO {
-  application?: string;
-  languagesCount?: number;
-  translationsCount?: number;
-}
-
