+++ broker-parent/broker-proxy/src/main/java/org/entcore/broker/proxy/I18nBrokerListener.java
@@ -4,8 +4,8 @@ import io.vertx.core.Future;
 import org.entcore.broker.api.BrokerListener;
 import org.entcore.broker.api.dto.i18n.FetchTranslationsRequestDTO;
 import org.entcore.broker.api.dto.i18n.FetchTranslationsResponseDTO;
-import org.entcore.broker.api.dto.i18n.RegisterI18nFilesRequestDTO;
-import org.entcore.broker.api.dto.i18n.RegisterI18nFilesResponseDTO;
+import org.entcore.broker.api.dto.i18n.RegisterTranslationFilesRequestDTO;
+import org.entcore.broker.api.dto.i18n.RegisterTranslationFilesResponseDTO;
 
 /**
  * This interface defines the methods that will be used to listen to events from the i18n broker.
@@ -17,7 +17,7 @@ public interface I18nBrokerListener {
    * @param request The request object containing the language and theme for which to fetch translations.
    * @return A response object containing a map of translation keys to their values in the requested language.
    */
-  @BrokerListener(subject = "{application}.i18n.translations.fetch", proxy = true)
+  @BrokerListener(subject = "i18n.{application}.fetch", proxy = true)
   Future<FetchTranslationsResponseDTO> fetchTranslations(FetchTranslationsRequestDTO request);
 
   /**
@@ -25,6 +25,6 @@ public interface I18nBrokerListener {
    * @param request The request object containing the details of the i18n files to register.
    * @return A response object indicating the success or failure of the registration process.
    */
-  @BrokerListener(subject = "{application}.i18n.translations.register", proxy = true)
-  Future<RegisterI18nFilesResponseDTO> registerI18nFiles(RegisterI18nFilesRequestDTO request);
+  @BrokerListener(subject = "i18n.{application}.register", proxy = true)
+  Future<RegisterTranslationFilesResponseDTO> registerI18nFiles(RegisterTranslationFilesRequestDTO request);
 }
