+++ infra/src/main/resources/public/js/viewers/xlsx.js/package.json
@@ -3,11 +3,11 @@
   "version": "2.3.1",
   "author": "Stephen Hardy",
   "description": "Convert data in base64 XLSX files into JavaScript objects - and back!",
-  "contributors": [ 
+  "contributors": [
     {
       "name": "Damien Feugas",
       "email": "damien.feugas@gmail.com"
-    } 
+    }
   ],
   "keywords": [
     "xls",
@@ -22,13 +22,13 @@
     "type": "git",
     "url": "https://github.com/stephen-hardy/xlsx.js"
   },
-  "engine": { 
+  "engine": {
     "node": ">=0.8.1"
   },
   "main": "xlsx.js",
   "dependencies": {
     "node-zip": "1.1.0"
-  },  
+  },
   "devDependencies": {
     "mocha": "1.8.2",
     "chai": "1.5.0",
